third_party/WebKit/LayoutTests/inspector/console/console-tainted-globals-expected.txt

Tests that overriding global methods (like Array.prototype.push, Math.max) will not break the inspector.


Running: evaluateInConsole

Running: testRuntimeAgentCallFunctionOn

Running: dumpConsoleMessages
testOverriddenArrayPushAndMathMax()
testOverriddenConstructorName()
testThrowConstructorName()
[1, 2, 3]
Object {constructor: Object}
Object {constructor: (...)}
